requirements.txt
pandas
PyPDF2
PyMuPDF
python-docx